+++ scipy/_lib/tests/meson.build
@@ -12,7 +12,6 @@ python_sources = [
   'test_import_cycles.py',
   'test_public_api.py',
   'test_scipy_version.py',
-  'test_tmpdirs.py',
   'test_warnings.py',
   'test_array_api.py',
 ]
