+++ scipy/io/tests/test_netcdf.py
@@ -6,7 +6,7 @@ import tempfile
 import warnings
 from io import BytesIO
 from glob import glob
-from contextlib import contextmanager
+from contextlib import chdir, contextmanager
 
 import numpy as np
 from numpy.testing import (assert_, assert_allclose, assert_equal,
@@ -15,7 +15,6 @@ import pytest
 from pytest import raises as assert_raises
 
 from scipy.io import netcdf_file
-from scipy._lib._tmpdirs import in_tempdir
 
 TEST_DATA_PATH = pjoin(dirname(__file__), 'data')
 
@@ -24,6 +23,39 @@ VARTYPE_EG = 'b'  # var type for example variable
 
 
 pytestmark = pytest.mark.thread_unsafe
+
+
+@contextmanager
+def in_tempdir():
+    ''' Create, return, and change directory to a temporary directory
+
+    Examples
+    --------
+    >>> import os
+    >>> my_cwd = os.getcwd()
+    >>> with in_tempdir() as tmpdir:
+    ...     _ = open('test.txt', 'wt').write('some text')
+    ...     assert os.path.isfile('test.txt')
+    ...     assert os.path.isfile(os.path.join(tmpdir, 'test.txt'))
+    >>> os.path.exists(tmpdir)
+    False
+    >>> os.getcwd() == my_cwd
+    True
+    '''
+    with tempfile.TemporaryDirectory() as td:
+        with chdir(td):
+            yield td
+
+
+def test_in_tempdir():
+    my_cwd = os.getcwd()
+    with in_tempdir() as tmpdir:
+        with open('test.txt', "w") as f:
+            f.write('some text')
+        assert os.path.isfile('test.txt')
+        assert os.path.isfile(os.path.join(tmpdir, 'test.txt'))
+    assert not os.path.exists(tmpdir)
+    assert os.getcwd() == my_cwd
 
 
 @contextmanager
